Update to OrientDB 1.5
Update the OrientDB engine to the version 1.5.0
Used the plocal engine which unfortunately does not manage transactions yet.
Maybe we have to consider to give the option to enable/disable transactions and to choose the storage type.
Keep in mind, however, that since OrientDB does not support transactions in cluster environments, this kind of feature will be disable when users migrate to the "as-a-service" service