Management of internal administrative user
Now there is a new user "internal_admin" used by the server to perform administrative tasks such as new user signup.

Commented out the use of the connection pool due a problem related to the password changes: connections in the pool continue to work with old password. We should open an issue on OrientDB