Adoption of the OrientDB plocal engine instead of the local one
The plocal engine should be more durable than the local one.
This should prevent the indices corruption that sometimes happens when the
BaasBox process is killed.

To migrate from previous version of baasbox, the administrator must
perform a backup (via console) on the older version and then perform a
restore on the new one.