New call_id parameter
New call_id value can be added to the querystring.
This can be any arbitrary string, and the server, if passed, will returned
it in the JSON reply.
This feature is useful in a async scenario when the client needs to know
what to do when the server responses come back in a different order from the
one with which they were sent