Image files resizing
Now the File of type "images" (i.e.png,jpg,bmp,gif,jpeg) can be resized
server side.

The clue is in the GET API:
GET /file/:id?sizeId=n where sizeId is a
zero-based index that points to one value defined into the ImageSetting
or
GET /file/:id/?resize=xxx
where xxx is one of the allowed resize value defined in the ImageSetting

Users belonging to the Administrator role can bypass the constraints and
require any type of resize.

Additional changes in this commit:

Fixed a bug that raised an Internal Server Error when the api
PUT /USER
missed in the payload the object visibleByTheUser

Changed the attribute visibleByFriend in visibleByFriends in Users'
profile

Fixed a bug into the Console that showed "undefined" in an alert message
when an invalid string pattern was supplied as ImageSetting value

Added the <=80px to the default values of allowed image resize patterns