Fixed reverse proxy error for Document list
Because absoluteURL() was being used to set the URL, using HTTPS and a
reverse proxy was causing errors when retrieving a list of documents. A
temporary fix could be using window.location.origin, but a better way
would be to utilize the configuration file and get the URL and port to
correctly set the URL (for Documents and any other places where
absoluteURL is being used).