Avoid to use Jackson to package the response to return to the clients.
If the JSON to return is huge (someone forgot to use pagination???), Jackson may fail to serialize it in a String causing an OutOfMemoryError in some cases.
Now we tried to avoid the use of Jackson at all, just building a valid JSON string by ourselves.

So say we all!