Workaround for issue #832
There is a problem with Nashorn when JSON objects are passed between differen ScriptObjectMirror objects.
This is the case when a plugin invokes a method exposed by another script via require/exports.
In particular this happens also with 'Box' methods because Box is implicitly imported by the plugin engine.

The workaround consists of to stringify the JSON object into the same ScriptObjectMirror object that generated it.
In this way the plugin receives a plain string that can be re-parsed into a new JSON that can be accessed and manipulated.