Implemented AppCode in URI path when a plugin is invoked
This is very helpful when you need to set up BaasBox plugin as a webhook in a third-party service.
Since every call to BaasBox must specify the AppCode, and this could be defined in a custom HTTP header, we needed a convenient way to pass this code in URL.
Now, if the AppCode is not found into the HTTP header, the framework will look the latest portion of the URI path, and try to use it as the AppCode.